Very basic teams implementation.

+++ Chemiklani.DAL/AppDbContext.cs
@@ -5,7 +5,7 @@ using Microsoft.AspNet.Identity.EntityFramework;
 
 namespace Chemiklani.DAL
 {
-    public class AppDbContext: IdentityDbContext<AppUser, AppRole, long, AppUserLogin, AppUserRole, AppUserClaim>
+    public class AppDbContext: IdentityDbContext<AppUser, AppRole, int, AppUserLogin, AppUserRole, AppUserClaim>
     {
 
         public AppDbContext() : base("DB")
