+++ Chemiklani.DAL/Entities/Task.cs
@@ -2,9 +2,9 @@ using Riganti.Utils.Infrastructure.Core;
 
 namespace Chemiklani.DAL.Entities
 {
-    public class Task: IEntity<long>
+    public class Task: IEntity<int>
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         public string Name { get; set; }
